@@ -2,7 +2,7 @@
   "name": "api",
   "private": true,
   "version": "0.1.0",
-  "description": "A Strapi application.",
+  "description": "Api Oath Hippocrate for data scientist.",
   "main": "./server.js",
   "scripts": {
     "setup": "cd admin && npm run setup",
@@ -12,16 +12,17 @@
     "postinstall": "node node_modules/strapi/lib/utils/post-install.js"
   },
   "devDependencies": {
-    "babel-eslint": "^7.1.1",
-    "eslint": "^3.12.2",
-    "eslint-config-airbnb": "^13.0.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-react": "^6.8.0"
+    "babel-eslint": "^8.2.3",
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-plugin-import": "^2.12.0",
+    "eslint-plugin-react": "^7.9.1"
   },
   "dependencies": {
     "lodash": "4.x.x",
-    "strapi": "3.0.0-alpha.12",
-    "strapi-mongoose": "3.0.0-alpha.12"
+    "npm-check-updates": "^2.14.2",
+    "strapi": "^3.0.0-alpha.12.5",
+    "strapi-mongoose": "^3.0.0-alpha.12.5"
   },
   "author": {
     "name": "stevensanseau",
