@@ -8,8 +8,6 @@ import React from 'react';
 import { defineMessages, FormattedMessage } from 'react-intl';
 import { PropTypes } from 'prop-types';
 
-import LocaleToggle from 'containers/LocaleToggle';
-
 import styles from './styles.scss';
 import messages from './messages.json';
 defineMessages(messages);
@@ -17,9 +15,10 @@ defineMessages(messages);
 function LeftMenuFooter({ version }) { // eslint-disable-line react/prefer-stateless-function
   return (
     <div className={styles.leftMenuFooter}>
-      <FormattedMessage {...messages.poweredBy} />
-      <a href="https://strapi.io" target="_blank">v{version}</a>
-      <LocaleToggle />
+      <div>
+        <FormattedMessage {...messages.poweredBy} />
+        <a href="https://strapi.io" target="_blank">v{version}</a>
+      </div>
     </div>
   );
 }
