@@ -19,10 +19,7 @@ import { get, includes, isFunction, map, omit } from 'lodash';
 import { compose } from 'redux';
 
 // Actions required for disabling and enabling the OverlayBlocker
-import {
-  disableGlobalOverlayBlocker,
-  enableGlobalOverlayBlocker,
-} from 'actions/overlayBlocker';
+import { disableGlobalOverlayBlocker, enableGlobalOverlayBlocker } from 'actions/overlayBlocker';
 
 import { pluginLoaded, updatePlugin } from 'containers/App/actions';
 import {
@@ -37,6 +34,8 @@ import { hideNotification } from 'containers/NotificationProvider/actions';
 // Design
 import ComingSoonPage from 'containers/ComingSoonPage';
 import Content from 'containers/Content';
+import LocaleToggle from 'containers/LocaleToggle';
+import CTAWrapper from 'components/CtaWrapper';
 import Header from 'components/Header/index';
 import HomePage from 'containers/HomePage/Loadable';
 import InstallPluginPage from 'containers/InstallPluginPage/Loadable';
@@ -59,17 +58,18 @@ import selectAdminPage from './selectors';
 
 import styles from './styles.scss';
 
-export class AdminPage extends React.Component { // eslint-disable-line react/prefer-stateless-function
+const PLUGINS_TO_BLOCK_PRODUCTION = ['content-type-builder', 'settings-manager'];
+
+export class AdminPage extends React.Component {
+  // eslint-disable-line react/prefer-stateless-function
   state = { hasAlreadyRegistereOtherPlugins: false };
 
-  getChildContext = () => (
-    {
-      disableGlobalOverlayBlocker: this.props.disableGlobalOverlayBlocker,
-      enableGlobalOverlayBlocker: this.props.enableGlobalOverlayBlocker,
-      plugins: this.props.plugins,
-      updatePlugin: this.props.updatePlugin,
-    }
-  );
+  getChildContext = () => ({
+    disableGlobalOverlayBlocker: this.props.disableGlobalOverlayBlocker,
+    enableGlobalOverlayBlocker: this.props.enableGlobalOverlayBlocker,
+    plugins: this.props.plugins,
+    updatePlugin: this.props.updatePlugin,
+  });
 
   componentDidMount() {
     this.checkLogin(this.props);
@@ -87,30 +87,52 @@ export class AdminPage extends React.Component { // eslint-disable-line react/pr
       }
     }
 
-    if (get(nextProps.plugins.toJS(), ['users-permissions', 'hasAdminUser']) !== get(this.props.plugins.toJS(), ['users-permissions', 'hasAdminUser'])) {
+    if (
+      get(nextProps.plugins.toJS(), ['users-permissions', 'hasAdminUser']) !==
+      get(this.props.plugins.toJS(), ['users-permissions', 'hasAdminUser'])
+    ) {
+      this.checkLogin(nextProps, true);
+    }
+
+    if (!this.hasUserPluginLoaded(this.props) && this.hasUserPluginLoaded(nextProps)) {
       this.checkLogin(nextProps);
     }
   }
 
-  checkLogin = (props) => {
+  checkLogin = (props, skipAction = false) => {
     if (props.hasUserPlugin && this.isUrlProtected(props) && !auth.getToken()) {
-      const endPoint = this.hasAdminUser(props) ? 'login': 'register';
+      if (!this.hasUserPluginLoaded(props)) {
+        return;
+      }
+
+      const endPoint = this.hasAdminUser(props) ? 'login' : 'register';
       this.props.history.push(`/plugins/users-permissions/auth/${endPoint}`);
     }
 
-    if (!this.isUrlProtected(props) && includes(props.location.pathname, 'auth/register') && this.hasAdminUser(props)) {
+    if (
+      !this.isUrlProtected(props) &&
+      includes(props.location.pathname, 'auth/register') &&
+      this.hasAdminUser(props) &&
+      !skipAction
+    ) {
       this.props.history.push('/plugins/users-permissions/auth/login');
     }
 
-    if (props.hasUserPlugin && !this.isUrlProtected(props) && !includes(props.location.pathname, 'auth/register') && !this.hasAdminUser(props)) {
+    if (
+      props.hasUserPlugin &&
+      !this.isUrlProtected(props) &&
+      !includes(props.location.pathname, 'auth/register') &&
+      !this.hasAdminUser(props)
+    ) {
       this.props.history.push('/plugins/users-permissions/auth/register');
     }
 
-    if (!props.hasUserPlugin || auth.getToken() && !this.state.hasAlreadyRegistereOtherPlugins) {
+    if (!props.hasUserPlugin || (auth.getToken() && !this.state.hasAlreadyRegistereOtherPlugins)) {
       map(omit(this.props.plugins.toJS(), ['users-permissions', 'email']), plugin => {
         switch (true) {
           case isFunction(plugin.bootstrap) && isFunction(plugin.pluginRequirements):
-            plugin.pluginRequirements(plugin)
+            plugin
+              .pluginRequirements(plugin)
               .then(plugin => {
                 return plugin.bootstrap(plugin);
               })
@@ -128,13 +150,35 @@ export class AdminPage extends React.Component { // eslint-disable-line react/pr
 
       this.setState({ hasAlreadyRegistereOtherPlugins: true });
     }
-  }
+  };
 
-  hasAdminUser = (props) => get(props.plugins.toJS(), ['users-permissions', 'hasAdminUser']);
+  hasUserPluginLoaded = props =>
+    typeof get(props.plugins.toJS(), ['users-permissions', 'hasAdminUser']) !== 'undefined';
 
-  isUrlProtected = (props) => !includes(props.location.pathname, get(props.plugins.toJS(), ['users-permissions', 'nonProtectedUrl']));
+  hasAdminUser = props => get(props.plugins.toJS(), ['users-permissions', 'hasAdminUser']);
+
+  isUrlProtected = props =>
+    !includes(props.location.pathname, get(props.plugins.toJS(), ['users-permissions', 'nonProtectedUrl']));
+
+  shouldDisplayLogout = () => auth.getToken() && this.props.hasUserPlugin && this.isUrlProtected(this.props);
 
   showLeftMenu = () => !includes(this.props.location.pathname, 'users-permissions/auth/');
+
+  retrievePlugins = () => {
+    const {
+      adminPage: { currentEnvironment },
+      plugins,
+    } = this.props;
+
+    if (currentEnvironment === 'production') {
+      let pluginsToDisplay = plugins;
+      PLUGINS_TO_BLOCK_PRODUCTION.map(plugin => (pluginsToDisplay = pluginsToDisplay.delete(plugin)));
+
+      return pluginsToDisplay;
+    }
+
+    return plugins;
+  };
 
   render() {
     const { adminPage } = this.props;
@@ -145,14 +189,15 @@ export class AdminPage extends React.Component { // eslint-disable-line react/pr
       <div className={styles.adminPage}>
         {this.showLeftMenu() && (
           <LeftMenu
-            plugins={this.props.plugins}
+            plugins={this.retrievePlugins()}
             layout={adminPage.layout}
             version={adminPage.strapiVersion}
           />
         )}
-        { auth.getToken() && this.props.hasUserPlugin && this.isUrlProtected(this.props) ? (
-          <Logout />
-        ) : ''}
+        <CTAWrapper>
+          {this.shouldDisplayLogout() && <Logout />}
+          <LocaleToggle isLogged={this.shouldDisplayLogout() === true} />
+        </CTAWrapper>
         <div className={styles.adminPageRightWrapper} style={style}>
           {header}
           <Content {...this.props} showLeftMenu={this.showLeftMenu()}>
@@ -216,13 +261,27 @@ const mapStateToProps = createStructuredSelector({
 
 function mapDispatchToProps(dispatch) {
   return {
-    disableGlobalOverlayBlocker: () => { dispatch(disableGlobalOverlayBlocker()); },
-    enableGlobalOverlayBlocker: () => { dispatch(enableGlobalOverlayBlocker()); },
-    getGaStatus: () => { dispatch(getGaStatus()); },
-    getLayout: () => { dispatch(getLayout()); },
-    onHideNotification: (id) => { dispatch(hideNotification(id)); },
-    pluginLoaded: (plugin) => { dispatch(pluginLoaded(plugin)); },
-    updatePlugin: (pluginId, updatedKey, updatedValue) => { dispatch(updatePlugin(pluginId, updatedKey, updatedValue)); },
+    disableGlobalOverlayBlocker: () => {
+      dispatch(disableGlobalOverlayBlocker());
+    },
+    enableGlobalOverlayBlocker: () => {
+      dispatch(enableGlobalOverlayBlocker());
+    },
+    getGaStatus: () => {
+      dispatch(getGaStatus());
+    },
+    getLayout: () => {
+      dispatch(getLayout());
+    },
+    onHideNotification: id => {
+      dispatch(hideNotification(id));
+    },
+    pluginLoaded: plugin => {
+      dispatch(pluginLoaded(plugin));
+    },
+    updatePlugin: (pluginId, updatedKey, updatedValue) => {
+      dispatch(updatePlugin(pluginId, updatedKey, updatedValue));
+    },
     dispatch,
   };
 }
@@ -231,10 +290,6 @@ const withConnect = connect(mapStateToProps, mapDispatchToProps);
 const withReducer = injectReducer({ key: 'adminPage', reducer });
 const withSaga = injectSaga({ key: 'adminPage', saga });
 
-export default compose(
-  withReducer,
-  withSaga,
-  withConnect,
-)(AdminPage);
+export default compose(withReducer, withSaga, withConnect)(AdminPage);
 
 // export default connect(mapStateToProps, mapDispatchToProps)(AdminPage);
