@@ -2,6 +2,7 @@ import { fork, call, put, takeLatest } from 'redux-saga/effects';
 import request from 'utils/request';
 
 import {
+  getCurrEnvSucceeded,
   getGaStatusSucceeded,
   getLayoutSucceeded,
   getStrapiVersionSucceeded,
@@ -10,11 +11,13 @@ import { GET_GA_STATUS, GET_LAYOUT } from './constants';
 
 function* getGaStatus() {
   try {
-    const [allowGa, strapiVersion] = yield [
+    const [allowGa, strapiVersion, currentEnvironment] = yield [
       call(request, '/admin/gaConfig', { method: 'GET' }),
       call(request, '/admin/strapiVersion', { method: 'GET' }),
+      call(request, '/admin/currentEnvironment', { method: 'GET' }),
     ];
 
+    yield put(getCurrEnvSucceeded(currentEnvironment.currentEnvironment));
     yield put(getGaStatusSucceeded(allowGa));
     yield put(getStrapiVersionSucceeded(strapiVersion.strapiVersion));
   } catch(err) {
