@@ -13,7 +13,7 @@ module.exports = {
     try {
       ctx.send({ currentEnvironment: strapi.app.env });
     } catch(err) {
-      ctx.badRequest(null, [{ messages: [{ id: 'An error occured' }] }]);
+      ctx.badRequest(null, [{ messages: [{ id: 'An error occurred' }] }]);
     }
   },
 
@@ -31,7 +31,7 @@ module.exports = {
       const allowGa = _.get(strapi.config, 'info.customs.allowGa', true);
       ctx.send({ allowGa });
     } catch(err) {
-      ctx.badRequest(null, [{ messages: [{ id: 'An error occured' }] }]);
+      ctx.badRequest(null, [{ messages: [{ id: 'An error occurred' }] }]);
     }
   },
 
@@ -41,7 +41,7 @@ module.exports = {
 
       return ctx.send({ layout });
     } catch(err) {
-      return ctx.badRequest(null, [{ messages: [{ id: 'An error occured' }] }]);
+      return ctx.badRequest(null, [{ messages: [{ id: 'An error occurred' }] }]);
     }
   },
 
@@ -54,14 +54,14 @@ module.exports = {
 
       strapi.log.info(`Installing ${plugin}...`);
 
-      exec(`node ${strapiBin} install ${plugin} ${port === '4000' ? '--dev' : ''}`);
+      exec(`node "${strapiBin}" install ${plugin} ${port === '4000' ? '--dev' : ''}`);
 
       ctx.send({ ok: true });
 
       strapi.reload();
     } catch(err) {
       strapi.reload.isWatching = true;
-      ctx.badRequest(null, [{ messages: [{ id: 'An error occured' }] }]);
+      ctx.badRequest(null, [{ messages: [{ id: 'An error occurred' }] }]);
     }
   },
 
@@ -75,7 +75,7 @@ module.exports = {
 
       ctx.send({ plugins });
     } catch(err) {
-      ctx.badRequest(null, [{ messages: [{ id: 'An error occured' }] }]);
+      ctx.badRequest(null, [{ messages: [{ id: 'An error occurred' }] }]);
     }
   },
 
@@ -87,14 +87,14 @@ module.exports = {
       strapi.reload.isWatching = false;
 
       strapi.log.info(`Uninstalling ${plugin}...`);
-      exec(`node ${strapiBin} uninstall ${plugin}`);
+      exec(`node "${strapiBin}" uninstall ${plugin}`);
 
       ctx.send({ ok: true });
 
       strapi.reload();
     } catch(err) {
       strapi.reload.isWatching = true;
-      ctx.badRequest(null, [{ messages: [{ id: 'An error occured' }] }]);
+      ctx.badRequest(null, [{ messages: [{ id: 'An error occurred' }] }]);
     }
   }
 };
