@@ -7,6 +7,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import CustomInputCheckbox from 'components/CustomInputCheckbox';
+
 import styles from './styles.scss';
 
 class TableHeader extends React.Component {
@@ -19,6 +21,18 @@ class TableHeader extends React.Component {
       this.props.onChangeSort(name);
     }
   }
+
+  renderBulk = () => (
+    <th key="bulk_action">
+      <CustomInputCheckbox
+        entriesToDelete={this.props.entriesToDelete}
+        isAll
+        name="all"
+        onChange={this.props.onClickSelectAll}
+        value={this.props.value}
+      />
+    </th>
+  );
 
   render() {
     // Generate headers list
@@ -47,23 +61,30 @@ class TableHeader extends React.Component {
     });
 
     // Add empty th for actions column.
-    headers.push(<th key={`th_action`}></th>);
+    headers.push(<th key="th_action"></th>);
 
     return (
       <thead className={styles.tableHeader}>
         <tr >
-          {headers}
+          {[this.renderBulk()].concat(headers)}
         </tr>
       </thead>
     );
   }
 }
 
+TableHeader.defaultProps = {
+  value: false,
+};
+
 TableHeader.propTypes = {
+  entriesToDelete: PropTypes.array.isRequired,
   headers: PropTypes.array.isRequired,
   onChangeSort: PropTypes.func.isRequired,
+  onClickSelectAll: PropTypes.func.isRequired,
   primaryKey: PropTypes.string.isRequired,
   sort: PropTypes.string.isRequired,
+  value: PropTypes.bool,
 };
 
 export default TableHeader;
