@@ -7,8 +7,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import moment from 'moment';
-import { isEmpty, isObject } from 'lodash';
+import { isEmpty, isObject, toString } from 'lodash';
 
+import CustomInputCheckbox from 'components/CustomInputCheckbox';
 import IcoContainer from 'components/IcoContainer';
 
 import styles from './styles.scss';
@@ -41,11 +42,15 @@ class TableRow extends React.Component {
       case 'decimal':
         return value && !isEmpty(value.toString()) ? value.toString() : '-';
       case 'boolean':
-        return value && !isEmpty(value.toString()) ? value.toString() : '-';
+        return value !== null ? toString(value) : '-';
       case 'date':
       case 'time':
       case 'datetime':
       case 'timestamp': {
+        if (value === null) {
+          return '-';
+        }
+
         const date = value && isObject(value) && value._isAMomentObject === true ?
           value :
           moment(value);
@@ -64,31 +69,51 @@ class TableRow extends React.Component {
     this.context.router.history.push(`${this.props.destination}${this.props.redirectUrl}`);
   }
 
+  renderAction = () => (
+    <td key='action' className={styles.actions}>
+      <IcoContainer
+        icons={[
+          { icoType: 'pencil', onClick: () => this.handleClick(this.props.destination) },
+          { id: this.props.record.id, icoType: 'trash', onClick: this.props.onDelete },
+        ]}
+      />
+    </td>
+  );
+
+  renderCells = () => {
+    const { headers } = this.props;
+    return [this.renderDelete()]
+      .concat(
+        headers.map((header, i) => (
+          <td key={i}>
+            <div className={styles.truncate}>
+              <div className={styles.truncated}>
+                {this.getDisplayedValue(
+                  header.type,
+                  this.props.record[header.name],
+                  header.name,
+                )}
+              </div>
+            </div>
+          </td>
+        )))
+      .concat([this.renderAction()]);
+  }
+
+  renderDelete = () => (
+    <td onClick={(e) => e.stopPropagation()} key="i">
+      <CustomInputCheckbox
+        name={this.props.record.id}
+        onChange={this.props.onChange}
+        value={this.props.value}
+      />
+    </td>
+  );
+
   render() {
-    // Generate cells
-    const cells = this.props.headers.map((header, i) => (
-      <td key={i}>
-        <div className={styles.truncate}>
-          <div className={styles.truncated}>
-            {this.getDisplayedValue(
-              header.type,
-              this.props.record[header.name],
-              header.name,
-            )}
-          </div>
-        </div>
-      </td>
-    ));
-
-    cells.push(
-      <td key='action' className={styles.actions}>
-        <IcoContainer icons={[{ icoType: 'pencil' }, { id: this.props.record.id, icoType: 'trash', onClick: this.props.onDelete }]} />
-      </td>
-    );
-
     return (
       <tr className={styles.tableRow} onClick={() => this.handleClick(this.props.destination)}>
-        {cells}
+        {this.renderCells()}
       </tr>
     );
   }
@@ -98,19 +123,22 @@ TableRow.contextTypes = {
   router: PropTypes.object.isRequired,
 };
 
+TableRow.defaultProps = {
+  value: false,
+};
+
 TableRow.propTypes = {
   destination: PropTypes.string.isRequired,
   headers: PropTypes.array.isRequired,
+  onChange: PropTypes.func.isRequired,
   onDelete: PropTypes.func,
   record: PropTypes.object.isRequired,
   redirectUrl: PropTypes.string.isRequired,
+  value: PropTypes.bool,
 };
 
 TableRow.defaultProps = {
   onDelete: () => {},
-  value: {
-    format: () => {},
-  },
 };
 
 export default TableRow;
