@@ -1,12 +1,12 @@
 const _ = require('lodash');
 
 module.exports = {
-  find: async function (params, populate) {
+  find: async function (params, populate, raw = false) {
     return this.query(function(qb) {
       _.forEach(params.where, (where, key) => {
-        if (_.isArray(where.value)) {
+        if (_.isArray(where.value) && where.symbol !== 'IN') {
           for (const value in where.value) {
-            qb[value ? 'where' : 'orWhere'](key, where.symbol, where.value[value])
+            qb[value ? 'where' : 'orWhere'](key, where.symbol, where.value[value]);
           }
         } else {
           qb.where(key, where.symbol, where.value);
@@ -14,7 +14,7 @@ module.exports = {
       });
 
       if (params.sort) {
-        qb.orderBy(params.sort);
+        qb.orderBy(params.sort.key, params.sort.order);
       }
 
       if (params.skip) {
@@ -26,16 +26,156 @@ module.exports = {
       }
     }).fetchAll({
       withRelated: populate || this.associations.map(x => x.alias)
-    });
+    }).then(data => raw ? data.toJSON() : data);
   },
 
-  count: async function (params) {
+  count: async function (params = {}) {
     return await this
       .forge()
+      .query(qb => {
+        _.forEach(params.where, (where, key) => {
+          if (_.isArray(where.value)) {
+            for (const value in where.value) {
+              qb[value ? 'where' : 'orWhere'](key, where.symbol, where.value[value]);
+            }
+          } else {
+            qb.where(key, where.symbol, where.value);
+          }
+        });
+      })
       .count();
   },
 
-  findOne: async function (params, populate, raw = true) {
+  search: async function (params, populate, raw = false) {
+    const associations = this.associations.map(x => x.alias);
+    const searchText = Object.keys(this._attributes)
+      .filter(attribute => attribute !== this.primaryKey && !associations.includes(attribute))
+      .filter(attribute => ['string', 'text'].includes(this._attributes[attribute].type));
+
+    const searchNoText = Object.keys(this._attributes)
+      .filter(attribute => attribute !== this.primaryKey && !associations.includes(attribute))
+      .filter(attribute => !['string', 'text', 'boolean', 'integer', 'decimal', 'float'].includes(this._attributes[attribute].type));
+
+    const searchInt = Object.keys(this._attributes)
+      .filter(attribute => attribute !== this.primaryKey && !associations.includes(attribute))
+      .filter(attribute => ['integer', 'decimal', 'float'].includes(this._attributes[attribute].type));
+
+    const searchBool = Object.keys(this._attributes)
+      .filter(attribute => attribute !== this.primaryKey && !associations.includes(attribute))
+      .filter(attribute => ['boolean'].includes(this._attributes[attribute].type));
+
+    const query = (params.search || '').replace(/[^a-zA-Z0-9.-\s]+/g, '');
+
+    return this.query(qb => {
+      // Search in columns which are not text value.
+      searchNoText.forEach(attribute => {
+        qb.orWhereRaw(`LOWER(${attribute}) LIKE '%${_.toLower(query)}%'`);
+      });
+
+      if (!_.isNaN(_.toNumber(query))) {
+        searchInt.forEach(attribute => {
+          qb.orWhereRaw(`${attribute} = ${_.toNumber(query)}`);
+        });
+      }
+
+      if (query === 'true' || query === 'false') {
+        searchBool.forEach(attribute => {
+          qb.orWhereRaw(`${attribute} = ${_.toNumber(query === 'true')}`);
+        });
+      }
+
+      // Search in columns with text using index.
+      switch (this.client) {
+        case 'pg': {
+          const searchQuery = searchText.map(attribute =>
+            _.toLower(attribute) === attribute
+              ? `to_tsvector(${attribute})`
+              : `to_tsvector('${attribute}')`
+          );
+
+          qb.orWhereRaw(`${searchQuery.join(' || ')} @@ to_tsquery(?)`, query);
+          break;
+        }
+        default:
+          qb.orWhereRaw(`MATCH(${searchText.join(',')}) AGAINST(? IN BOOLEAN MODE)`, `*${query}*`);
+          break;
+      }
+
+      if (params.sort) {
+        qb.orderBy(params.sort.key, params.sort.order);
+      }
+
+      if (params.skip) {
+        qb.offset(_.toNumber(params.skip));
+      }
+
+      if (params.limit) {
+        qb.limit(_.toNumber(params.limit));
+      }
+    }).fetchAll({
+      width: populate || associations
+    }).then(data => raw ? data.toJSON() : data);
+  },
+
+  countSearch: async function (params = {}) {
+    const associations = this.associations.map(x => x.alias);
+    const searchText = Object.keys(this._attributes)
+      .filter(attribute => attribute !== this.primaryKey && !associations.includes(attribute))
+      .filter(attribute => ['string', 'text'].includes(this._attributes[attribute].type));
+
+    const searchNoText = Object.keys(this._attributes)
+      .filter(attribute => attribute !== this.primaryKey && !associations.includes(attribute))
+      .filter(attribute => !['string', 'text', 'boolean', 'integer', 'decimal', 'float'].includes(this._attributes[attribute].type));
+
+    const searchInt = Object.keys(this._attributes)
+      .filter(attribute => attribute !== this.primaryKey && !associations.includes(attribute))
+      .filter(attribute => ['integer', 'decimal', 'float'].includes(this._attributes[attribute].type));
+
+    const searchBool = Object.keys(this._attributes)
+      .filter(attribute => attribute !== this.primaryKey && !associations.includes(attribute))
+      .filter(attribute => ['boolean'].includes(this._attributes[attribute].type));
+
+    const query = (params.search || '').replace(/[^a-zA-Z0-9.-\s]+/g, '');
+
+
+    return this.query(qb => {
+      // Search in columns which are not text value.
+      searchNoText.forEach(attribute => {
+        qb.orWhereRaw(`LOWER(${attribute}) LIKE '%${_.toLower(query)}%'`);
+      });
+
+      if (!_.isNaN(_.toNumber(query))) {
+        searchInt.forEach(attribute => {
+          qb.orWhereRaw(`${attribute} = ${_.toNumber(query)}`);
+        });
+      }
+
+      if (query === 'true' || query === 'false') {
+        searchBool.forEach(attribute => {
+          qb.orWhereRaw(`${attribute} = ${_.toNumber(query === 'true')}`);
+        });
+      }
+
+      // Search in columns with text using index.
+      switch (this.client) {
+        case 'pg': {
+          const searchQuery = searchText.map(attribute =>
+            _.toLower(attribute) === attribute
+              ? `to_tsvector(${attribute})`
+              : `to_tsvector('${attribute}')`
+          );
+
+          qb.orWhereRaw(`${searchQuery.join(' || ')} @@ to_tsquery(?)`, query);
+          break;
+        }
+        default:
+          qb.orWhereRaw(`MATCH(${searchText.join(',')}) AGAINST(? IN BOOLEAN MODE)`, `*${query}*`);
+          break;
+      }
+    }).count();
+  },
+
+  findOne: async function (params, populate) {
     const record = await this
       .forge({
         [this.primaryKey]: params[this.primaryKey]
@@ -44,18 +184,18 @@ module.exports = {
         withRelated: populate || this.associations.map(x => x.alias)
       });
 
-    const data = record ? record.toJSON() : record;
+    const data = record.toJSON ? record.toJSON() : record;
 
     // Retrieve data manually.
     if (_.isEmpty(populate)) {
       const arrayOfPromises = this.associations
         .filter(association => ['manyMorphToOne', 'manyMorphToMany'].includes(association.nature))
-        .map(association => {
+        .map(association => { // eslint-disable-line no-unused-vars
           return this.morph.forge()
             .where({
               [`${this.collectionName}_id`]: params[this.primaryKey]
             })
-            .fetchAll()
+            .fetchAll();
         });
 
       const related = await Promise.all(arrayOfPromises);
@@ -82,7 +222,7 @@ module.exports = {
       .forge(values)
       .save()
       .catch((err) => {
-        if (err.detail) {
+        if (err.detail) {
           const field = _.last(_.words(err.detail.split('=')[0]));
           err = { message: `This ${field} is already taken`, field };
         }
@@ -106,223 +246,9 @@ module.exports = {
   },
 
   update: async function (params) {
-    const virtualFields = [];
-    const response = await module.exports.findOne.call(this, params);
-
-    // Only update fields which are on this document.
-    const values = params.parseRelationships === false ? params.values : Object.keys(JSON.parse(JSON.stringify(params.values))).reduce((acc, current) => {
-      const association = this.associations.filter(x => x.alias === current)[0];
-      const details = this._attributes[current];
-
-      if (_.get(this._attributes, `${current}.isVirtual`) !== true && _.isUndefined(association)) {
-        acc[current] = params.values[current];
-      } else {
-        switch (association.nature) {
-          case 'oneWay':
-            acc[current] = _.get(params.values[current], this.primaryKey, params.values[current]) || null;
-
-            break;
-          case 'oneToOne':
-            if (response[current] !== params.values[current]) {
-              const value = _.isNull(params.values[current]) ? response[current] : params.values;
-              const recordId = _.isNull(params.values[current]) ? value[this.primaryKey] || value.id || value._id : value[current];
-
-              if (response[current] && _.isObject(response[current]) && response[current][this.primaryKey] !== value[current]) {
-                virtualFields.push(
-                  strapi.query(details.collection || details.model, details.plugin).update({
-                    id: response[current][this.primaryKey],
-                    values: {
-                      [details.via]: null
-                    },
-                    parseRelationships: false
-                  })
-                );
-              }
-
-              // Remove previous relationship asynchronously if it exists.
-              virtualFields.push(
-                strapi.query(details.model || details.collection, details.plugin).findOne({ id : recordId })
-                  .then(record => {
-                    if (record && _.isObject(record[details.via]) && record[details.via][current] !== value[current]) {
-                      return module.exports.update.call(this, {
-                        id: record[details.via][this.primaryKey] || record[details.via].id,
-                        values: {
-                          [current]: null
-                        },
-                        parseRelationships: false
-                      });
-                    }
-
-                    return Promise.resolve();
-                  })
-              );
-
-              // Update the record on the other side.
-              // When params.values[current] is null this means that we are removing the relation.
-              virtualFields.push(strapi.query(details.model || details.collection, details.plugin).update({
-                id: recordId,
-                values: {
-                  [details.via]: _.isNull(params.values[current]) ? null : value[this.primaryKey] || value.id || value._id
-                },
-                parseRelationships: false
-              }));
-
-              acc[current] = _.isNull(params.values[current]) ? null : value[current];
-            }
-
-            break;
-          case 'oneToMany':
-          case 'manyToOne':
-          case 'manyToMany':
-            if (response[current] && _.isArray(response[current]) && current !== 'id') {
-              // Records to add in the relation.
-              const toAdd = _.differenceWith(params.values[current], response[current], (a, b) =>
-                a[this.primaryKey].toString() === b[this.primaryKey].toString()
-              );
-              // Records to remove in the relation.
-              const toRemove = _.differenceWith(response[current], params.values[current], (a, b) =>
-                a[this.primaryKey].toString() === b[this.primaryKey].toString()
-              )
-                .filter(x => toAdd.find(y => x.id === y.id) === undefined);
-
-              // Push the work into the flow process.
-              toAdd.forEach(value => {
-                value[details.via] = params.values[this.primaryKey] || params[this.primaryKey];
-
-                virtualFields.push(strapi.query(details.model || details.collection, details.plugin).addRelation({
-                  id: value[this.primaryKey] || value.id || value._id,
-                  values: association.nature === 'manyToMany' ? params.values : value,
-                  foreignKey: current
-                }));
-              });
-
-              toRemove.forEach(value => {
-                value[details.via] = null;
-
-                virtualFields.push(strapi.query(details.model || details.collection, details.plugin).removeRelation({
-                  id: value[this.primaryKey] || value.id || value._id,
-                  values: association.nature === 'manyToMany' ? params.values : value,
-                  foreignKey: current
-                }));
-              });
-            } else if (_.get(this._attributes, `${current}.isVirtual`) !== true) {
-              acc[current] = params.values[current];
-            }
-
-            break;
-          case 'manyMorphToMany':
-          case 'manyMorphToOne':
-            // Update the relational array.
-            params.values[current].forEach(obj => {
-              const model = obj.source && obj.source !== 'content-manager' ?
-                strapi.plugins[obj.source].models[obj.ref]:
-                strapi.models[obj.ref];
-
-              // Remove existing relationship because only one file
-              // can be related to this field.
-              if (association.nature === 'manyMorphToOne') {
-                virtualFields.push(
-                  module.exports.removeRelationMorph.call(this, {
-                    alias: association.alias,
-                    ref: model.collectionName,
-                    refId: obj.refId,
-                    field: obj.field
-                  })
-                  .then(() =>
-                    module.exports.addRelationMorph.call(this, {
-                      id: response[this.primaryKey],
-                      alias: association.alias,
-                      ref: model.collectionName,
-                      refId: obj.refId,
-                      field: obj.field
-                    })
-                  )
-                );
-              } else {
-                virtualFields.push(module.exports.addRelationMorph.call(this, {
-                  id: response[this.primaryKey],
-                  alias: association.alias,
-                  ref: model.collectionName,
-                  refId: obj.refId,
-                  field: obj.field
-                }));
-              }
-            });
-            break;
-          case 'oneToManyMorph':
-          case 'manyToManyMorph':
-            const transformToArrayID = (array) => {
-              if(_.isArray(array)) {
-                return array.map(value => {
-                  if (_.isPlainObject(value)) {
-                    return value._id || value.id;
-                  }
-
-                  return value;
-                })
-              }
-
-              if (association.type === 'model' || (association.type === 'collection' && _.isObject(array))) {
-                return _.isEmpty(_.toString(array)) ? [] : transformToArrayID([array]);
-              }
-
-              return [];
-            };
-
-            // Compare array of ID to find deleted files.
-            const currentValue = transformToArrayID(response[current]).map(id => id.toString());
-            const storedValue = transformToArrayID(params.values[current]).map(id => id.toString());
-
-            const toAdd = _.difference(storedValue, currentValue);
-            const toRemove = _.difference(currentValue, storedValue);
-
-            toAdd.forEach(id => {
-              virtualFields.push(strapi.query(details.model || details.collection, details.plugin).addRelationMorph({
-                id,
-                alias: association.via,
-                ref: this.collectionName,
-                refId: response.id,
-                field: association.alias
-              }));
-            });
-
-            // Update the relational array.
-            toRemove.forEach(id => {
-              virtualFields.push(strapi.query(details.model || details.collection, details.plugin).removeRelationMorph({
-                id,
-                alias: association.via,
-                ref: this.collectionName,
-                refId: response.id,
-                field: association.alias
-              }));
-            });
-            break;
-          case 'oneMorphToOne':
-          case 'oneMorphToMany':
-            break;
-          default:
-        }
-      }
-
-      return acc;
-    }, {});
-
-    if (!_.isEmpty(values)) {
-      virtualFields.push(this
-        .forge({
-          [this.primaryKey]: params[this.primaryKey]
-        })
-        .save(values, {
-          patch: true
-        }));
-    } else {
-      virtualFields.push(Promise.resolve(_.assign(response, params.values)));
-    }
-
-    // Update virtuals fields.
-    const process = await Promise.all(virtualFields);
-
-    return process[process.length - 1];
+    // Call the business logic located in the hook.
+    // This function updates no-relational and relational data.
+    return this.updateRelations(params);
   },
 
   delete: async function (params) {
@@ -333,87 +259,11 @@ module.exports = {
       .destroy();
   },
 
-  addRelation: async function (params) {
-    const association = this.associations.filter(x => x.via === params.foreignKey)[0];
-
-    if (!association) {
-      // Resolve silently.
-      return Promise.resolve();
-    }
-
-    switch (association.nature) {
-      case 'oneToOne':
-      case 'oneToMany':
-      case 'manyToOne':
-        return module.exports.update.call(this, params);
-      case 'manyToMany':
-        return this.forge({
-          [this.primaryKey]: params[this.primaryKey]
-        })[association.alias]().attach(params.values[this.primaryKey]);
-      default:
-        // Resolve silently.
-        return Promise.resolve();
-    }
-  },
-
-  removeRelation: async function (params) {
-    const association = this.associations.filter(x => x.via === params.foreignKey)[0];
-
-    if (!association) {
-      // Resolve silently.
-      return Promise.resolve();
-    }
-
-    switch (association.nature) {
-      case 'oneToOne':
-      case 'oneToMany':
-      case 'manyToOne':
-        return module.exports.update.call(this, params);
-      case 'manyToMany':
-        return this.forge({
-          [this.primaryKey]: params[this.primaryKey]
-        })[association.alias]().detach(params.values[this.primaryKey]);
-      default:
-        // Resolve silently.
-        return Promise.resolve();
-    }
-  },
-
-  addRelationMorph: async function (params) {
-    const record = await this.morph.forge()
-      .where({
-        [`${this.collectionName}_id`]: params.id,
-        [`${params.alias}_id`]: params.refId,
-        [`${params.alias}_type`]: params.ref,
-        field: params.field
+  deleteMany: async function (params) {
+    return await this
+      .query(function(qb) {
+        return qb.whereIn('id', params.id);
       })
-      .fetch({
-        withRelated: this.associations.map(x => x.alias)
-      });
-
-    const entry = record ? record.toJSON() : record;
-
-    if (entry) {
-      return Promise.resolve();
-    }
-
-    return await this.morph.forge({
-        [`${this.collectionName}_id`]: params.id,
-        [`${params.alias}_id`]: params.refId,
-        [`${params.alias}_type`]: params.ref,
-        field: params.field
-      })
-      .save();
-  },
-
-  removeRelationMorph: async function (params) {
-    return await this.morph.forge()
-      .where(_.omitBy({
-        [`${this.collectionName}_id`]: params.id,
-        [`${params.alias}_id`]: params.refId,
-        [`${params.alias}_type`]: params.ref,
-        field: params.field
-      }, _.isEmpty))
       .destroy();
   }
 };
