@@ -44,14 +44,22 @@ module.exports = {
   },
 
   find: async ctx => {
+    // Search
+    if (!_.isEmpty(ctx.request.query._q)) {
+      ctx.body = await strapi.plugins['content-manager'].services['contentmanager'].search(ctx.params, ctx.request.query);
+
+      return;
+    }
+
+    // Default list with filters or not.
     ctx.body = await strapi.plugins['content-manager'].services['contentmanager'].fetchAll(ctx.params, ctx.request.query);
   },
 
   count: async ctx => {
-    const { source } = ctx.request.query;
-
-    // Count using `queries` system
-    const count = await strapi.plugins['content-manager'].services['contentmanager'].count(ctx.params, source);
+    // Search
+    const count = !_.isEmpty(ctx.request.query._q)
+      ? await strapi.plugins['content-manager'].services['contentmanager'].countSearch(ctx.params, ctx.request.query)
+      : await strapi.plugins['content-manager'].services['contentmanager'].count(ctx.params, ctx.request.query);
 
     ctx.body = {
       count: _.isNumber(count) ? count : _.toNumber(count)
@@ -100,4 +108,8 @@ module.exports = {
   delete: async ctx => {
     ctx.body = await strapi.plugins['content-manager'].services['contentmanager'].delete(ctx.params, ctx.request.query);
   },
+
+  deleteAll: async ctx => {
+    ctx.body = await strapi.plugins['content-manager'].services['contentmanager'].deleteMany(ctx.params, ctx.request.query);
+  }
 };
