@@ -8,20 +8,44 @@ const _ = require('lodash');
 
 module.exports = {
   fetchAll: async (params, query) => {
-    const { limit, skip = 0, sort, query : request, queryAttribute, source, page, populate = [] } = query;
+    const { limit, skip, sort, query : request, queryAttribute, source, page, populate = [] } = query; // eslint-disable-line no-unused-vars
+    const filters = strapi.utils.models.convertParams(params.model, query);
+    const where = !_.isEmpty(request) ? request : filters.where;
 
     // Find entries using `queries` system
     return await strapi.query(params.model, source).find({
-      limit,
-      skip,
-      sort,
-      where: request,
+      limit: limit || filters.limit,
+      skip: skip || filters.start || 0,
+      sort: sort || filters.sort,
+      where,
       queryAttribute,
     }, populate);
   },
 
-  count: async (params, source) => {
-    return await strapi.query(params.model, source).count();
+  search: async (params, query) => {
+    const { limit, skip, sort, source, _q, populate = [] } = query; // eslint-disable-line no-unused-vars
+    const filters = strapi.utils.models.convertParams(params.model, query);
+
+    // Find entries using `queries` system
+    return await strapi.query(params.model, source).search({
+      limit: limit || filters.limit,
+      skip: skip || filters.start || 0,
+      sort: sort || filters.sort,
+      search: _q
+    }, populate);
+  },
+
+  countSearch: async (params, query) => {
+    const { source, _q } = query;
+
+    return await strapi.query(params.model, source).countSearch({ search: _q });
+  },
+
+  count: async (params, query) => {
+    const { source } = query;
+    const filters = strapi.utils.models.convertParams(params.model, query);
+
+    return await strapi.query(params.model, source).count({ where: filters.where });
   },
 
   fetch: async (params, source, populate, raw = true) => {
@@ -126,10 +150,13 @@ module.exports = {
   },
 
   delete: async (params, { source }) => {
-    const response = await strapi.query(params.model, source).findOne({
+    const query = strapi.query(params.model, source);
+    const primaryKey = query.primaryKey;
+    const response = await query.findOne({
       id: params.id
     });
 
+    params[primaryKey] = response[primaryKey];
     params.values = Object.keys(JSON.parse(JSON.stringify(response))).reduce((acc, current) => {
       const association = (strapi.models[params.model] || strapi.plugins[source].models[params.model]).associations.filter(x => x.alias === current)[0];
 
@@ -151,4 +178,53 @@ module.exports = {
       id: params.id
     });
   },
+
+  deleteMany: async (params, query) => {
+    const { source } = query;
+    const { model } = params;
+    const primaryKey = strapi.query(model, source).primaryKey;
+    const toRemove = Object.keys(query).reduce((acc, curr) => {
+      if (curr !== 'source') {
+        return acc.concat([query[curr]]);
+      }
+
+      return acc;
+    }, []);
+
+    const filters = strapi.utils.models.convertParams(model, { [`${primaryKey}_in`]: toRemove });
+    const entries = await strapi.query(model, source).find({ where: filters.where }, null, true);
+    const associations = strapi.query(model, source).associations;
+
+    for (let i = 0; i < entries.length; ++i) {
+      const entry = entries[i];
+
+      associations.forEach(association => {
+        if (entry[association.alias]) {
+          switch (association.nature) {
+            case 'oneWay':
+            case 'oneToOne':
+            case 'manyToOne':
+            case 'oneToManyMorph':
+              entry[association.alias] = null;
+              break;
+            case 'oneToMany':
+            case 'manyToMany':
+            case 'manyToManyMorph':
+              entry[association.alias] = [];
+              break;
+            default:
+          }
+        }
+      });
+
+      await strapi.query(model, source).update({
+        [primaryKey]: entry[primaryKey],
+        values: _.pick(entry, associations.map(a => a.alias))
+      });
+    }
+
+    return strapi.query(model, source).deleteMany({
+      [primaryKey]: toRemove,
+    });
+  }
 };
