@@ -22,9 +22,12 @@ import IcoPassword from '../../assets/images/icon_password.png';
 import IcoRelation from '../../assets/images/icon_relation.png';
 import IcoString from '../../assets/images/icon_string.png';
 import IcoText from '../../assets/images/icon_text.png';
+import IcoEnum from '../../assets/images/icon_enum.png';
 import styles from './styles.scss';
 
 /* eslint-disable jsx-a11y/no-static-element-interactions */
+/* eslint-disable jsx-a11y/click-events-have-key-events */
+/* eslint-disable react/jsx-curly-brace-presence */
 class AttributeRow extends React.Component {
   // eslint-disable-line react/prefer-stateless-function
   constructor(props) {
@@ -43,8 +46,7 @@ class AttributeRow extends React.Component {
       decimal: IcoNumber,
       email: IcoEmail,
       password: IcoPassword,
-      // TODO add Enumeration icon
-      enumeration: IcoJson,
+      enumeration: IcoEnum,
     };
     this.state = {
       showWarning: false,
