@@ -48,6 +48,10 @@ class PopUpForm extends React.Component { // eslint-disable-line react/prefer-st
     );
   }
 
+  handleSubmit = (e) => {
+    this.props.onSubmit(e, true);
+  }
+
   renderInput = (item, key) => {
     // const customBootstrapClass = 'col-md-6'
     let customBootstrapClass = item.type === 'textarea' ?
@@ -123,23 +127,29 @@ class PopUpForm extends React.Component { // eslint-disable-line react/prefer-st
     return <FormattedMessage id={this.props.popUpTitle} />;
   }
 
+  renderFooter = () => {
+    const { popUpFormType, buttonSubmitMessage, toggle, noButtons, onSubmit } = this.props;
+    const handleToggle = toggle;
+
+    if (noButtons) {
+      return <div className={styles.modalFooter} />;
+    }
+
+    return (
+      <ModalFooter className={styles.modalFooter}>
+        <Button onClick={handleToggle} className={styles.secondary}><FormattedMessage id="content-type-builder.form.button.cancel" /></Button>
+        {popUpFormType !== 'contentType' && <Button type="submit" onClick={this.handleSubmit} className={styles.primaryAddShape}><FormattedMessage id="content-type-builder.button.attributes.add" /></Button>}
+        <Button type="submit" onClick={onSubmit} className={styles.primary}><FormattedMessage id={`content-type-builder.${buttonSubmitMessage}`} /></Button>{' '}
+      </ModalFooter>
+    );
+  }
+
   render() {
     const navContainer = this.props.noNav ? '' : this.renderNavContainer();
     const modalBodyStyle = this.props.renderModalBody ? { paddingTop: '2.3rem' } : {};
     const modalBody = this.props.renderModalBody ? this.props.renderModalBody()
       : map(this.props.form.items, (item, key ) => this.renderInput(item, key));
-
-    const loader = this.props.showLoader ?
-      <Button onClick={this.props.onSubmit} type="submit" className={styles.primary} disabled={this.props.showLoader}><p className={styles.saving}><span>.</span><span>.</span><span>.</span></p></Button>
-      : <Button type="submit" onClick={this.props.onSubmit} className={styles.primary}><FormattedMessage id={`content-type-builder.${this.props.buttonSubmitMessage}`} /></Button>;
-
-    const handleToggle = this.props.toggle;
-    const modalFooter = this.props.noButtons ? <div className={styles.modalFooter} />
-      : <ModalFooter className={styles.modalFooter}>
-        <Button onClick={handleToggle} className={styles.secondary}><FormattedMessage id="content-type-builder.form.button.cancel" /></Button>
-        {loader}{' '}
-      </ModalFooter>;
-
+    
     return (
       <div className={styles.popUpForm}>
         <Modal isOpen={this.props.isOpen} toggle={this.props.toggle} className={`${styles.modalPosition}`}>
@@ -153,13 +163,13 @@ class PopUpForm extends React.Component { // eslint-disable-line react/prefer-st
           <ModalBody className={styles.modalBody} style={modalBodyStyle}>
             <form onSubmit={this.props.onSubmit}>
               <div className="container-fluid">
-                <div className={`row ${this.props.renderModalBody ? 'justify-content-center' : ''}`}>
+                <div className="row">
                   {modalBody}
                 </div>
               </div>
             </form>
           </ModalBody>
-          {modalFooter}
+          {this.renderFooter()}
         </Modal>
       </div>
     );
@@ -200,6 +210,7 @@ PopUpForm.propTypes = {
     PropTypes.bool,
     PropTypes.func,
   ]),
+  popUpFormType: PropTypes.string,
   popUpHeaderNavLinks: PropTypes.array,
   popUpTitle: PropTypes.string.isRequired,
   renderCustomPopUpHeader: PropTypes.oneOfType([
@@ -213,7 +224,6 @@ PopUpForm.propTypes = {
   ]).isRequired,
   routePath: PropTypes.string,
   selectOptions: PropTypes.array,
-  showLoader: PropTypes.bool,
   toggle: PropTypes.func.isRequired,
   values: PropTypes.object,
 };
@@ -229,11 +239,11 @@ PopUpForm.defaultProps = {
   overrideHandleBlurCondition: false,
   overrideRenderInput: false,
   overrideRenderInputCondition: false,
+  popUpFormType: '',
   popUpHeaderNavLinks: [],
   renderCustomPopUpHeader: false,
   routePath: '',
   selectOptions: [],
-  showLoader: false,
   values: {},
 };
 
