@@ -1,13 +1,14 @@
 /**
-*
-* PopUpRelations
-*
-*/
+ *
+ * PopUpRelations
+ *
+ */
 
 import React from 'react';
 import PropTypes from 'prop-types';
 import { findIndex, get, isEmpty, map, take, takeRight } from 'lodash';
 import { FormattedMessage } from 'react-intl';
+import pluralize from 'pluralize';
 
 import { Button, Modal, ModalHeader, ModalBody, ModalFooter } from 'reactstrap';
 import Input from 'components/InputsIndex';
@@ -17,12 +18,21 @@ import RelationNaturePicker from 'components/RelationNaturePicker';
 import styles from './styles.scss';
 
 /* eslint-disable jsx-a11y/tabindex-no-positive */
-class PopUpRelations extends React.Component { // eslint-disable-line react/prefer-stateless-function
+class PopUpRelations extends React.Component {
+  // eslint-disable-line react/prefer-stateless-function
   constructor(props) {
     super(props);
     this.popUpHeaderNavLinks = [
-      { name: 'defineRelation', message: 'content-type-builder.popUpForm.navContainer.relation', nameToReplace: 'advancedSettings' },
-      { name: 'advancedSettings', message: 'content-type-builder.popUpForm.navContainer.advanced', nameToReplace: 'defineRelation' },
+      {
+        name: 'defineRelation',
+        message: 'content-type-builder.popUpForm.navContainer.relation',
+        nameToReplace: 'advancedSettings',
+      },
+      {
+        name: 'advancedSettings',
+        message: 'content-type-builder.popUpForm.navContainer.advanced',
+        nameToReplace: 'defineRelation',
+      },
     ];
   }
 
@@ -33,12 +43,93 @@ class PopUpRelations extends React.Component { // eslint-disable-line react/pref
   }
 
   componentWillReceiveProps(nextProps) {
-    if (isEmpty(this.props.dropDownItems) && !isEmpty(nextProps.dropDownItems) && !this.props.isEditting) {
+    if (
+      isEmpty(this.props.dropDownItems) &&
+      !isEmpty(nextProps.dropDownItems) &&
+      !this.props.isEditting
+    ) {
       this.init(nextProps);
     }
   }
 
-  init = (props) => {
+  setPlaceholders = (firstCTName, secondCTName, relationType, values = this.props.values) => {
+    switch (relationType) {
+      case 'oneToMany':
+        firstCTName = pluralize(firstCTName);
+        break;
+      case 'manyToOne':
+        secondCTName = pluralize(secondCTName);
+        break;
+      case 'manyToMany':
+        firstCTName = pluralize(firstCTName);
+        secondCTName = pluralize(secondCTName);
+        break;
+      default:
+        // Do nothing
+    }
+
+    if (get(this.props.contentType, 'name') !== get(values, 'params.target')) {
+      this.props.onChange({ target: { name: 'name', value: firstCTName } });
+      this.props.onChange({ target: { name: 'params.key', value: secondCTName } });
+      this.props.resetFormErrors();
+    } else {
+      this.props.onChange({ target: { name: 'name', value: '' } });
+      this.props.onChange({ target: { name: 'params.key', value: '' } });
+    }
+  }
+
+  handleChange = e => {
+    this.props.onChange(e);
+    const shouldResetKeyParams = e.target.value === 'oneWay';
+
+    if (!this.props.isEditting && !shouldResetKeyParams) {
+      this.setPlaceholders(
+        get(this.props.values, ['params', 'target']),
+        get(this.props.contentType, 'name'),
+        e.target.value,
+      );
+    }
+
+    if (shouldResetKeyParams) {
+      this.props.onChange({ target: { name: 'params.key', value: '-' } });
+    }
+  }
+
+  handleClick = e => {
+    const value = e.target.id.split('.');
+    [
+      {
+        target: {
+          type: 'string',
+          value: value[0],
+          name: 'params.target',
+        },
+      },
+      {
+        target: {
+          type: 'string',
+          value: value[1] !== 'undefined' ? value[1] : '',
+          name: 'params.pluginValue',
+        },
+      },
+    ].map(target => this.props.onChange(target));
+
+    if (!this.props.isEditting) {
+      if (get(this.props.contentType, 'name') !== value[0]) {
+        this.setPlaceholders(
+          value[0],
+          get(this.props.contentType, 'name'),
+          get(this.props.values, ['params', 'nature']),
+          value[0],
+        );
+      } else {
+        this.props.onChange({ target: { name: 'name', value: '' } });
+        this.props.onChange({ target: { name: 'params.key', value: '' } });
+      }
+    }
+  };
+
+  init = props => {
     const target = {
       name: 'params.target',
       type: 'string',
@@ -56,7 +147,14 @@ class PopUpRelations extends React.Component { // eslint-disable-line react/pref
         },
       });
     }
-  }
+
+    if (get(props.contentType, 'name') !== get(props.dropDownItems, ['0', 'name'])) {
+      [
+        { target: { name: 'name', value: get(props.dropDownItems, ['0', 'name']) } },
+        { target: { name: 'params.key', value: get(props.contentType, 'name') } },
+      ].map(target => this.props.onChange(target));
+    }
+  };
 
   renderNavContainer = () => (
     <div className={styles.navContainer}>
@@ -70,7 +168,7 @@ class PopUpRelations extends React.Component { // eslint-disable-line react/pref
         />
       ))}
     </div>
-  )
+  );
 
   renderModalBodyAdvanced = () => (
     <ModalBody className={`${styles.modalBodyAdvanced}`}>
@@ -95,7 +193,10 @@ class PopUpRelations extends React.Component { // eslint-disable-line react/pref
         <div className={styles.inputContainer}>
           <div className="row">
             {map(takeRight(this.props.form.items, 2), (value, index) => {
-              const addon = index === 0 ? get(this.props.values, 'name') : get(this.props.values, ['params', 'key']);
+              const addon =
+                index === 0
+                  ? get(this.props.values, 'name')
+                  : get(this.props.values, ['params', 'key']);
               return (
                 <Input
                   key={index}
@@ -117,12 +218,25 @@ class PopUpRelations extends React.Component { // eslint-disable-line react/pref
         </div>
       </div>
     </ModalBody>
-  )
+  );
 
   renderModalBodyRelations = () => {
-    const header = get(this.props.values, ['params', 'pluginValue']) ?
-      get(this.props.dropDownItems, [findIndex(this.props.dropDownItems, {'name': get(this.props.values, ['params', 'target']), source: get(this.props.values, ['params', 'pluginValue']) })])
-      : get(this.props.dropDownItems, [findIndex(this.props.dropDownItems, ['name', get(this.props.values, ['params', 'target'])])]);
+    const header = get(this.props.values, ['params', 'pluginValue'])
+      ? get(this.props.dropDownItems, [
+        findIndex(this.props.dropDownItems, {
+          name: get(this.props.values, ['params', 'target']),
+          source: get(this.props.values, ['params', 'pluginValue']),
+        }),
+      ])
+      : get(this.props.dropDownItems, [
+        findIndex(this.props.dropDownItems, [
+          'name',
+          get(this.props.values, ['params', 'target']),
+        ]),
+      ]);
+
+    const errs = findIndex(this.props.formErrors, ['name',get(this.props.form, ['items', '0', 'name'])]) !== -1 ? this.props.formErrors[findIndex(this.props.formErrors, ['name', get(this.props.form, ['items', '0', 'name'])])].errors: [];
+    const errors = findIndex(this.props.formErrors, ['name', get(this.props.form, ['items', '1', 'name'])]) !== -1 ? this.props.formErrors[findIndex(this.props.formErrors, ['name', get(this.props.form, ['items', '1', 'name'])])].errors : [];
 
     return (
       <ModalBody className={`${styles.modalBody} ${styles.flex}`}>
@@ -138,11 +252,11 @@ class PopUpRelations extends React.Component { // eslint-disable-line react/pref
           onSubmit={this.props.onSubmit}
           onChange={this.props.onChange}
           didCheckErrors={this.props.didCheckErrors}
-          errors={findIndex(this.props.formErrors, ['name', get(this.props.form, ['items', '0', 'name'])]) !== -1 ? this.props.formErrors[findIndex(this.props.formErrors, ['name', get(this.props.form, ['items', '0', 'name'])])].errors : []}
+          errors={errs}
         />
         <RelationNaturePicker
           selectedIco={get(this.props.values, ['params', 'nature'])}
-          onChange={this.props.onChange}
+          onChange={this.handleChange}
           contentTypeName={get(this.props.contentType, 'name')}
           contentTypeTarget={get(this.props.values, ['params', 'target'])}
         />
@@ -156,24 +270,27 @@ class PopUpRelations extends React.Component { // eslint-disable-line react/pref
           value={get(this.props.values, ['params', 'key'])}
           onChange={this.props.onChange}
           didCheckErrors={this.props.didCheckErrors}
-          errors={findIndex(this.props.formErrors, ['name', get(this.props.form, ['items', '1', 'name'])]) !== -1 ? this.props.formErrors[findIndex(this.props.formErrors, ['name', get(this.props.form, ['items', '1', 'name'])])].errors : []}
+          errors={errors}
           dropDownItems={this.props.dropDownItems}
+          onClick={this.handleClick}
         />
       </ModalBody>
     );
-  }
+  };
 
   render() {
-    const loader = this.props.showLoader ?
-      <Button onClick={this.props.onSubmit} type="submit" className={styles.primary} disabled={this.props.showLoader}><p className={styles.saving}><span>.</span><span>.</span><span>.</span></p></Button>
-      : <Button type="submit" onClick={this.props.onSubmit} className={styles.primary}><FormattedMessage id="content-type-builder.form.button.continue" /></Button>;
-
-    const modalBody = this.props.showRelation ? this.renderModalBodyRelations():  this.renderModalBodyAdvanced();
+    const modalBody = this.props.showRelation
+      ? this.renderModalBodyRelations()
+      : this.renderModalBodyAdvanced();
     const handleToggle = this.props.toggle;
 
     return (
       <div className={styles.popUpRelations}>
-        <Modal isOpen={this.props.isOpen} toggle={this.props.toggle} className={`${styles.modalPosition}`}>
+        <Modal
+          isOpen={this.props.isOpen}
+          toggle={this.props.toggle}
+          className={`${styles.modalPosition}`}
+        >
           <ModalHeader toggle={this.props.toggle} className={styles.popUpFormHeader} />
           <div className={styles.headerContainer}>
             <div className={styles.titleContainer}>
@@ -190,8 +307,13 @@ class PopUpRelations extends React.Component { // eslint-disable-line react/pref
           {modalBody}
 
           <ModalFooter className={styles.modalFooter}>
-            <Button onClick={handleToggle} className={styles.secondary}><FormattedMessage id="content-type-builder.form.button.cancel" /></Button>
-            {loader}{' '}
+            <Button onClick={handleToggle} className={styles.secondary}>
+              <FormattedMessage id="content-type-builder.form.button.cancel" />
+            </Button>
+            <Button type="submit" onClick={(e) => this.props.onSubmit(e, true)} className={styles.primaryAddShape}><FormattedMessage id="content-type-builder.button.attributes.add" /></Button>
+            <Button type="submit" onClick={this.props.onSubmit} className={styles.primary}>
+              <FormattedMessage id="content-type-builder.form.button.continue" />
+            </Button>{' '}
           </ModalFooter>
         </Modal>
       </div>
@@ -203,19 +325,14 @@ PopUpRelations.propTypes = {
   contentType: PropTypes.object,
   didCheckErrors: PropTypes.bool.isRequired,
   dropDownItems: PropTypes.array,
-  form: PropTypes.oneOfType([
-    PropTypes.array.isRequired,
-    PropTypes.object.isRequired,
-  ]).isRequired,
-  formErrors: PropTypes.oneOfType([
-    PropTypes.array,
-    PropTypes.object,
-  ]).isRequired,
+  form: PropTypes.oneOfType([PropTypes.array.isRequired, PropTypes.object.isRequired]).isRequired,
+  formErrors: PropTypes.oneOfType([PropTypes.array, PropTypes.object]).isRequired,
   isEditting: PropTypes.bool,
   isOpen: PropTypes.bool.isRequired,
   onChange: PropTypes.func.isRequired,
   onSubmit: PropTypes.func.isRequired,
   popUpTitle: PropTypes.string.isRequired,
+  resetFormErrors: PropTypes.func.isRequired,
   routePath: PropTypes.string.isRequired,
   showLoader: PropTypes.bool,
   showRelation: PropTypes.bool.isRequired,
