@@ -23,7 +23,6 @@ import ContentHeader from 'components/ContentHeader';
 import EmptyAttributesView from 'components/EmptyAttributesView';
 import Form from 'containers/Form';
 import List from 'components/List';
-import NoTableWarning from 'components/NoTableWarning';
 import PluginLeftMenu from 'components/PluginLeftMenu';
 
 import forms from 'containers/Form/forms.json';
@@ -35,7 +34,6 @@ import { storeData } from '../../utils/storeData';
 
 import {
   cancelChanges,
-  checkIfTableExists,
   deleteAttribute,
   modelFetch,
   modelFetchSucceeded,
@@ -54,6 +52,7 @@ availableAttributes.push('integer', 'decimal', 'float');
 
 /* eslint-disable jsx-a11y/no-static-element-interactions */
 /* eslint-disable react/jsx-wrap-multilines */
+/* eslint-disable react/jsx-curly-brace-presence */
 
 export class ModelPage extends React.Component { // eslint-disable-line react/prefer-stateless-function
   constructor(props) {
@@ -83,10 +82,6 @@ export class ModelPage extends React.Component { // eslint-disable-line react/pr
       if (this.state.contentTypeTemporary && storeData.getContentType()) {
         this.props.modelFetchSucceeded({ model: storeData.getContentType() });
       }
-    }
-
-    if (this.props.modelPage.didFetchModel !== nextProps.modelPage.didFetchModel) {
-      this.props.checkIfTableExists();
     }
   }
 
@@ -151,9 +146,12 @@ export class ModelPage extends React.Component { // eslint-disable-line react/pr
   }
 
   handleDelete = (attributeName) => {
-    const index = findIndex(this.props.modelPage.model.attributes, ['name', attributeName]);
-    const parallelAttributeIndex = findIndex(this.props.modelPage.model.attributes, (attr) => attr.params.key === attributeName);
-
+    const { modelPage: { model } } = this.props;
+    const index = findIndex(model.attributes, ['name', attributeName]);
+    const attributeToRemove = get(model, ['attributes', index]);
+    const parallelAttributeIndex = attributeToRemove.name === attributeToRemove.params.key ?
+      -1 : findIndex(model.attributes, (attr) => attr.params.key === attributeName);
+      
     this.props.deleteAttribute(index, this.props.match.params.modelName, parallelAttributeIndex !== -1);
   }
 
@@ -163,7 +161,7 @@ export class ModelPage extends React.Component { // eslint-disable-line react/pr
 
     // Display a notification if the attribute is not present in the ones that the ctb handles
     if (!has(attribute.params, 'nature') && !includes(availableAttributes, attribute.params.type)) {
-      return strapi.notification.info('content-type-builder.notification.info.enumeration');
+      return strapi.notification.info('content-type-builder.notification.info.disable');
     }
     const settingsType = attribute.params.type ? 'baseSettings' : 'defineRelation';
     const parallelAttributeIndex = findIndex(this.props.modelPage.model.attributes, ['name', attribute.params.key]);
@@ -265,7 +263,6 @@ export class ModelPage extends React.Component { // eslint-disable-line react/pr
     // Url to redirects the user if he modifies the temporary content type name
     const redirectRoute = replace(this.props.match.path, '/:modelName', '');
     const addButtons  = get(storeData.getContentType(), 'name') === this.props.match.params.modelName && size(get(storeData.getContentType(), 'attributes')) > 0 || this.props.modelPage.showButtons;
-    const showNoTableWarning = this.props.modelPage.tableExists ? '' : <NoTableWarning modelName={this.props.modelPage.model.name} />;
     const contentHeaderDescription = this.props.modelPage.model.description || 'content-type-builder.modelPage.contentHeader.emptyDescription.description';
     const content = size(this.props.modelPage.model.attributes) === 0 ?
       <EmptyAttributesView onClickAddAttribute={this.handleClickAddAttribute} /> :
@@ -303,7 +300,6 @@ export class ModelPage extends React.Component { // eslint-disable-line react/pr
 
                 />
                 {content}
-                {showNoTableWarning}
               </div>
             </div>
           </div>
@@ -332,9 +328,7 @@ ModelPage.contextTypes = {
 
 ModelPage.propTypes = {
   cancelChanges: PropTypes.func.isRequired,
-  checkIfTableExists: PropTypes.func.isRequired,
   deleteAttribute: PropTypes.func.isRequired,
-  // history: PropTypes.object.isRequired,
   location: PropTypes.object.isRequired,
   match: PropTypes.object.isRequired,
   menu: PropTypes.array.isRequired,
@@ -356,7 +350,6 @@ function mapDispatchToProps(dispatch) {
   return bindActionCreators(
     {
       cancelChanges,
-      checkIfTableExists,
       deleteAttribute,
       modelFetch,
       modelFetchSucceeded,
