@@ -1,6 +1,6 @@
 'use strict';
 
-const path = require('path')
+const path = require('path');
 const fs = require('fs');
 const _ = require('lodash');
 const generator = require('strapi-generate');
@@ -105,13 +105,14 @@ module.exports = {
     return {
       name: _.get(model, 'info.name', 'model.name.missing'),
       description: _.get(model, 'info.description', 'model.description.missing'),
+      mainField: _.get(model, 'info.mainField', ''),
       connection: model.connection,
       collectionName: model.collectionName,
       attributes: attributes
     };
   },
 
-  getConnections: () => {
+  getConnections: () => {
     return _.keys(strapi.config.currentEnvironment.database.connections);
   },
 
@@ -125,7 +126,7 @@ module.exports = {
         rootPath: strapi.config.appPath,
         args: {
           api: name,
-          description: _.replace(description, /\"/g, '\\"'),
+          description: _.replace(description, /\"/g, '\\"'), // eslint-disable-line no-useless-escape
           attributes,
           connection,
           collectionName: !_.isEmpty(collectionName) ? collectionName : undefined,
@@ -190,7 +191,7 @@ module.exports = {
             [attribute.params.multiple ? 'collection' : 'model']: 'file',
             via,
             plugin: 'upload'
-          }
+          };
         }
       } else if (_.has(attribute, 'params.target')) {
         const relation = attribute.params;
@@ -201,6 +202,7 @@ module.exports = {
         };
 
         switch (relation.nature) {
+          case 'oneWay':
           case 'oneToOne':
           case 'manyToOne':
             attr.model = relation.target;
@@ -212,11 +214,13 @@ module.exports = {
           default:
         }
 
-        attr.via = relation.key;
+        if(relation.nature !== 'oneWay') {
+          attr.via = relation.key;
+        }
         attr.dominant = relation.dominant;
 
-        if (relation.pluginValue) {
-          attr.plugin = relation.pluginValue;
+        if (_.trim(relation.pluginValue)) {
+          attr.plugin = _.trim(relation.pluginValue);
         }
 
         attrs[attribute.name] = attr;
@@ -353,25 +357,32 @@ module.exports = {
             const attr = {};
 
             switch (params.nature) {
+              case 'oneWay':
+                return;
               case 'oneToOne':
               case 'oneToMany':
                 attr.model = model.toLowerCase();
                 break;
               case 'manyToOne':
                 attr.collection = model.toLowerCase();
-              case 'manyToMany':
+                break;
+              case 'manyToMany': {
                 attr.collection = model.toLowerCase();
 
                 if (!params.dominant) {
                   attr.dominant = true;
                 }
                 break;
+              }
               default:
             }
 
             attr.via = name;
             attr.columnName = params.targetColumnName;
-            attr.plugin = source;
+
+            if (_.trim(source)) {
+              attr.plugin = _.trim(source);
+            }
 
             modelJSON.attributes[params.key] = attr;
 
@@ -454,7 +465,7 @@ module.exports = {
           }
         }
       }
-    }
+    };
 
     const recurciveDeleteFiles = folderPath => {
       try {
@@ -490,7 +501,7 @@ module.exports = {
           }
         });
       }
-    }
+    };
 
     recurciveDeleteFiles(apiPath);
 
