@@ -8,11 +8,13 @@
 const _ = require('lodash');
 const path = require('path');
 const glob = require('glob');
-const { graphqlKoa, graphiqlKoa } = require('apollo-server-koa');
+const { graphqlKoa } = require('apollo-server-koa');
+const koaPlayground = require('graphql-playground-middleware-koa').default;
+const depthLimit = require('graphql-depth-limit');
 
 module.exports = strapi => {
   return {
-    beforeInitialize: async function() {
+    beforeInitialize: async function(){
       // Try to inject this hook just after the others hooks to skip the router processing.
       if (!_.get(strapi.config.hook.load, 'after')) {
         _.set(strapi.config.hook.load, 'after', []);
@@ -68,14 +70,14 @@ module.exports = strapi => {
        */
 
       // Set path with initial state.
-      _.set(strapi.plugins.graphql, 'config._schema.graphql', { definition: ``, query: ``, type : {}, resolver: {} });
+      _.set(strapi.plugins.graphql, 'config._schema.graphql', { definition: '', query: '', type : {}, resolver: {} });
 
       // Merge user API.
       Object.keys(strapi.api || {}).reduce((acc, current) => {
         const { definition, query, type, resolver } = _.get(strapi.api[current], 'config.schema.graphql', {});
 
-        acc.definition += definition || ``;
-        acc.query += query || ``;
+        acc.definition += definition || '';
+        acc.query += query || '';
 
         return _.merge(acc, {
           type,
@@ -87,8 +89,8 @@ module.exports = strapi => {
       Object.keys(strapi.plugins || {}).reduce((acc, current) => {
         const { definition, query, type, resolver } = _.get(strapi.plugins[current], 'config.schema.graphql', {});
 
-        acc.definition += definition || ``;
-        acc.query += query || ``;
+        acc.definition += definition || '';
+        acc.query += query || '';
 
         return _.merge(acc, {
           type,
@@ -101,19 +103,28 @@ module.exports = strapi => {
       const schema = strapi.plugins.graphql.services.graphql.generateSchema();
 
       if (_.isEmpty(schema)) {
-        strapi.log.warn(`GraphQL schema has not been generated because it's empty`);
+        strapi.log.warn('GraphQL schema has not been generated because it\'s empty');
 
         return cb();
       }
 
       const router = strapi.koaMiddlewares.routerJoi();
 
-      router.post(strapi.plugins.graphql.config.endpoint, async (ctx, next) => graphqlKoa({ schema, context: ctx })(ctx, next));
-      router.get(strapi.plugins.graphql.config.endpoint, async (ctx, next) => graphqlKoa({ schema, context: ctx })(ctx, next));
+      router.post(strapi.plugins.graphql.config.endpoint, async (ctx, next) => graphqlKoa({
+        schema,
+        context: ctx,
+        validationRules: [ depthLimit(strapi.plugins.graphql.config.depthLimit) ]
+      })(ctx, next));
 
-      // Disable GraphiQL in production environment.
+      router.get(strapi.plugins.graphql.config.endpoint, async (ctx, next) => graphqlKoa({
+        schema,
+        context: ctx,
+        validationRules: [ depthLimit(strapi.plugins.graphql.config.depthLimit) ]
+      })(ctx, next));
+
+      // Disable GraphQL Playground in production environment.
       if (strapi.config.environment !== 'production') {
-        router.get('/graphiql', graphiqlKoa({ endpointURL: strapi.plugins.graphql.config.endpoint }));
+        router.get('/playground', koaPlayground({ endpoint: strapi.plugins.graphql.config.endpoint}));
       }
 
       strapi.app.use(router.middleware());
