@@ -70,17 +70,25 @@ module.exports = {
    */
 
   create: async (ctx) => {
-    if ((await strapi.store({
+    const advanced = await strapi.store({
       environment: '',
       type: 'plugin',
       name: 'users-permissions',
       key: 'advanced'
-    }).get()).unique_email && ctx.request.body.email) {
+    }).get();
+
+    if (advanced.unique_email && ctx.request.body.email) {
       const user = await strapi.query('user', 'users-permissions').findOne({ email: ctx.request.body.email });
 
       if (user) {
         return ctx.badRequest(null, ctx.request.admin ? [{ messages: [{ id: 'Auth.form.error.email.taken', field: ['email'] }] }] : 'Email is already taken.');
       }
+    }
+
+    if (!ctx.request.body.role) {
+      const defaultRole = await strapi.query('role', 'users-permissions').findOne({ type: advanced.default_role }, []);
+
+      ctx.request.body.role = defaultRole._id || defaultRole.id;
     }
 
     try {
@@ -98,7 +106,7 @@ module.exports = {
    * @return {Object}
    */
 
-  update: async (ctx, next) => {
+  update: async (ctx) => {
     try {
       const advancedConfigs = await strapi.store({
         environment: '',
@@ -117,7 +125,7 @@ module.exports = {
 
       const user = await strapi.plugins['users-permissions'].services.user.fetch(ctx.params);
 
-      if (_.get(ctx.request, 'body.password') === user.password) {
+      if (_.get(ctx.request, 'body.password') === user.password) {
         delete ctx.request.body.password;
       }
 
@@ -150,8 +158,15 @@ module.exports = {
    * @return {Object}
    */
 
-  destroy: async (ctx, next) => {
+  destroy: async (ctx) => {
     const data = await strapi.plugins['users-permissions'].services.user.remove(ctx.params);
+
+    // Send 200 `ok`
+    ctx.send(data);
+  },
+
+  destroyAll: async (ctx) => {
+    const data = await strapi.plugins['users-permissions'].services.user.removeAll(ctx.params, ctx.request.query);
 
     // Send 200 `ok`
     ctx.send(data);
