@@ -83,7 +83,7 @@ module.exports = {
         resolve(null);
       } else {
         bcrypt.hash(user.password, 10, (err, hash) => {
-          resolve(hash)
+          resolve(hash);
         });
       }
     });
@@ -113,6 +113,30 @@ module.exports = {
     }
 
     return strapi.query('user', 'users-permissions').delete(params);
+  },
+
+  removeAll: async (params, query) => {
+    // Use Content Manager business logic to handle relation.
+    if (strapi.plugins['content-manager']) {
+      params.model = 'user';
+      query.source = 'users-permissions';
+
+      return await strapi.plugins['content-manager'].services['contentmanager'].deleteMany(params, query);
+    }
+
+    // TODO remove this logic when we develop plugins' dependencies
+    const primaryKey = strapi.query('user', 'users-permissions').primaryKey;
+    const toRemove = Object.keys(query).reduce((acc, curr) => {
+      if (curr !== 'source') {
+        return acc.concat([query[curr]]);
+      }
+
+      return acc;
+    }, []);
+
+    return strapi.query('user', 'users-permissions').deleteMany({
+      [primaryKey]: toRemove,
+    });
   },
 
   validatePassword: (password, hash) => {
